Added pages for registry lists and updates samples
